1.05 Renamed Button package to Panel. WindowCM class moved to package Panel.
