--- src/com/CalculatingMachine/Buttons/BuildOfButtons.java
@@ -1,4 +1,0 @@
-package com.CalculatingMachine.Buttons;
-
-public class BuildOfButtons {
-}
