+++ src/com/CalculatingMachine/Panel/BuildOfButtons.java
@@ -1,0 +1,4 @@
+package com.CalculatingMachine.Panel;
+
+public class BuildOfButtons {
+}
